Remove `SentenceInformationService` and move `getSentenceAndOffenceDetails` to `PersonService`
I think keeping person/prisoner related information in one service makes sense.  Will be adding future methods for getting a persons main offence as well as any other offences.
