+++ server/controllers/shared/index.ts
@@ -9,7 +9,6 @@ const controllers = (services: Services) => {
     services.organisationService,
     services.personService,
     services.referralService,
-    services.sentenceInformationService,
   )
 
   return {
