+++ server/services/index.ts
@@ -5,7 +5,6 @@ import healthCheck from './healthCheck'
 import OrganisationService from './organisationService'
 import PersonService from './personService'
 import ReferralService from './referralService'
-import SentenceInformationService from './sentenceInformationService'
 import UserService from './userService'
 import {
   caseloadClientBuilder,
@@ -20,9 +19,8 @@ import {
 
 const services = () => {
   const organisationService = new OrganisationService(prisonClientBuilder)
-  const personService = new PersonService(hmppsAuthClientBuilder, prisonerClientBuilder)
+  const personService = new PersonService(hmppsAuthClientBuilder, prisonApiClientBuilder, prisonerClientBuilder)
   const referralService = new ReferralService(referralClientBuilder)
-  const sentenceInformationService = new SentenceInformationService(prisonApiClientBuilder)
   const userService = new UserService(hmppsManageUsersClientBuilder, caseloadClientBuilder)
   const courseService = new CourseService(courseClientBuilder, userService)
 
@@ -31,22 +29,12 @@ const services = () => {
     organisationService,
     personService,
     referralService,
-    sentenceInformationService,
     userService,
   }
 }
 
 type Services = ReturnType<typeof services>
 
-export {
-  CourseService,
-  OrganisationService,
-  PersonService,
-  ReferralService,
-  SentenceInformationService,
-  UserService,
-  healthCheck,
-  services,
-}
+export { CourseService, OrganisationService, PersonService, ReferralService, UserService, healthCheck, services }
 
 export type { Services }
