+++ server/services/personService.ts
@@ -1,14 +1,21 @@
 import createError from 'http-errors'
 import type { ResponseError } from 'superagent'
 
-import type { HmppsAuthClient, PrisonerClient, RestClientBuilder, RestClientBuilderWithoutToken } from '../data'
+import type {
+  HmppsAuthClient,
+  PrisonApiClient,
+  PrisonerClient,
+  RestClientBuilder,
+  RestClientBuilderWithoutToken,
+} from '../data'
 import { PersonUtils } from '../utils'
 import type { Person } from '@accredited-programmes/models'
-import type { Caseload } from '@prison-api'
+import type { Caseload, SentenceAndOffenceDetails } from '@prison-api'
 
 export default class PersonService {
   constructor(
     private readonly hmppsAuthClientBuilder: RestClientBuilderWithoutToken<HmppsAuthClient>,
+    private readonly prisonApiClientBuilder: RestClientBuilder<PrisonApiClient>,
     private readonly prisonerClientBuilder: RestClientBuilder<PrisonerClient>,
   ) {}
 
@@ -41,6 +48,29 @@ export default class PersonService {
       throw createError(knownError.status || 500, knownError, {
         userMessage: `Error fetching prisoner ${prisonNumber}.`,
       })
+    }
+  }
+
+  async getSentenceAndOffenceDetails(
+    token: Express.User['token'],
+    bookingId: Person['bookingId'],
+  ): Promise<SentenceAndOffenceDetails> {
+    try {
+      const prisonApiClient = this.prisonApiClientBuilder(token)
+      return await prisonApiClient.findSentenceAndOffenceDetails(bookingId)
+    } catch (error) {
+      const knownError = error as ResponseError
+
+      if (knownError.status === 404) {
+        throw createError(knownError.status, `Sentence and offence details for booking with ID ${bookingId} not found.`)
+      }
+
+      const errorMessage =
+        knownError.message === 'Internal Server Error'
+          ? `Error fetching sentence and offence details with booking ID ${bookingId}.`
+          : knownError.message
+
+      throw createError(knownError.status || 500, errorMessage)
     }
   }
 }
